# business/openapi.yaml
openapi: 3.1.0
info:
  title: rust-invoice-system — business API
  version: "1.0.0"
  description: |
    Invoicing, charging and event delivery.

    ## Two audiences, one port

    Which endpoints need a token is decided by **who is asking**, not by whether the request
    writes. `GET /invoices` is gated and `POST /invoices/{id}/pay` is not, because raising or
    voiding an invoice is the business's own affair, while whoever is being billed has no token
    and cannot be given one.

    - **Operator** — the business's back office. Requires `X-API-Token`.
    - **Payer** — the person being billed. `POST /invoices/{id}/pay` and
      `GET /payment_intents/{id}`, both open by necessity. Neither exposes anything about an
      invoice beyond the charge the caller is already making.
    - **Docker** — `GET /health`, open so the container healthcheck can reach it.

    An unauthenticated request to an operator endpoint gets `401 unauthorized`; a *wrong* token
    gets exactly the same, since telling that caller their header was well formed and merely
    incorrect hands them the one bit they were missing. A path matching no route stays `404`
    rather than `401`, which would reveal which paths exist.

    ## Error format

    Every failure answers with the same shape, so a caller parses failures one way:

    ```json
    { "error": "invoice_not_payable" }
    ```

    One addition, on `POST /invoices/{id}/pay`: once a charge has reached the payment service, every
...

...

...

...
servers:
...
paths:
...
  /invoices/{id}:
    get:
      tags: [Invoices]
      summary: Read one invoice
      operationId: getInvoice
      security: [{ operatorToken: [] }]
      parameters: [{ $ref: "#/components/parameters/InvoiceId" }]
      responses:
        "200":
          description: The invoice.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Invoice" }
        "400":
          description: "`invalid_invoice_id`"
          $ref: "#/components/responses/Error"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404":
          description: "`invoice_not_found`"
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
...
  /payment_intents/{id}:
    get:
      tags: [Payments]
      summary: What became of one charge attempt
      operationId: getPaymentIntent
      security: []
      description: |
        **No `X-API-Token`** — the other half of `POST /invoices/{id}/pay`, for a caller that got
        a `504` and has no idea whether the card was charged.

        **The recorded row is the answer.** No call to the payment service, no settlement, no
        side effect of any kind — that row is what the rest of the system acted on. The cost is
        that `pending` means *this service does not know yet*, not *the card was not charged*; a
        charge whose response was lost reads `pending` until the daily reconciler settles it.
      parameters:
        - name: id
          in: path
          required: true
          description: The payment service's own id.
          schema: { type: string }
          example: pi_55e3caceeccd43fa80cde32d642ec100
      responses:
        "200":
          description: The recorded status.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/PaymentIntent" }
        "404":
          description: "`payment_intent_not_found` — this service has no record of that intent."
          $ref: "#/components/responses/Error"
        "500": { $ref: "#/components/responses/Internal" }
...
components:
  securitySchemes:
    operatorToken:
      type: apiKey
      in: header
      name: X-API-Token
      description: |
        One shared static token, compared in constant time. A missing token and a wrong one get
        the same `401`.
  parameters:
...
    InvoiceId:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
  responses:
    Error:
      description: A failure, in the standard shape.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: "`unauthorized` — no token, or the wrong one."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: unauthorized }
    Internal:
      description: "`internal_error`"
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
          example: { error: internal_error }
...
  schemas:
    Error:
      type: object
      description: The shape every failure answers with, bar the two documented exceptions.
      properties:
        error:
          type: string
          description: A stable machine-readable code. Never prose.
          example: invoice_not_payable
      required: [error]
...
    InvoiceState:
      type: string
      enum: [draft, ready, processing, processed, void]
      description: |
        `draft` — being prepared, editable.
        `ready` — finalized and payable.
        `processing` — a charge is in flight; no operator endpoint can move it.
        `processed` — paid. Terminal.
        `void` — written off. Terminal.
...
    Invoice:
      type: object
      properties:
        id: { type: string, format: uuid }
        customer_id: { type: string, format: uuid }
        total_cents:
          type: integer
          minimum: 0
          description: Minor units, summed from the line items by the server.
        due_date: { type: string, format: date }
        state: { $ref: "#/components/schemas/InvoiceState" }
      required: [id, customer_id, total_cents, due_date, state]
...
    PaymentStatus:
      type: string
      enum: [pending, succeeded, failed]
      description: |
        `pending` means **this service does not know yet**, not that the card was not charged.
    PaymentIntent:
      type: object
      properties:
        payment_intent_id: { type: string }
        status: { $ref: "#/components/schemas/PaymentStatus" }
      required: [payment_intent_id, status]
...